+++ CHANGELOG.md
@@ -22,6 +22,7 @@ All notable changes to this project will be documented in this file.
 - The `amqp_1` input and output should no longer spam logs with timeout errors during graceful termination.
 - Fixed a potential crash when the `contains` bloblang method was used to compare complex types.
 - Fixed an issue where the `kafka_franz` input or output wouldn't use TLS connections without custom certificate configuration.
+- Fixed structural cycle in the CUE representation of the `retry` output.
 
 ### Changed